backend/requirements_ml.txt
--extra-index-url https://download.pytorch.org/whl/cu124
# API Dependencies
alembic
asyncio
asyncpg
fastapi[standard]
httpx==0.28.1
passlib==1.7.4
pydantic==2.12.3
pydantic_settings==2.11.0
PyJWT==2.9.0
pytest==8.3.5
pytest_asyncio==0.21.1
redis==7.0.0
SQLAlchemy==2.0.44

# ML Dependencies
playwright
chromadb==1.2.1
docling==2.58.0
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.30
langchain_text_splitters==0.3.9
langgraph==0.6.6
openai==1.99.9
python-dotenv==1.1.1
Requests==2.32.5
sentence_transformers==5.1.0
exa-py
torch
pysqlite3-binary
git+https://github.com/Skeletonboi/yt-transcript-util.git